backend/requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
apturl==0.5.2
argcomplete==3.1.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
ayatana-settings==23.11.14
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
deja-dup-caja==0.0.9
distro==1.9.0
distro-info==1.7+build1
Django==4.2.11
django-extensions==3.2.3
duplicity==2.1.4
fasteners==0.18
folder-color-caja==0.0.86
folder-color-common==0.0.86
frozenlist==1.4.0
gpg==1.18.0
httplib2==0.20.4
idna==3.6
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kazam==1.4.5
keyring==24.3.1
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libvirt-python==10.0.0
louis==3.29.0
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mate-hud==22.10.3
mate-tweak==22.10.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
multidict==6.0.4
mutagen==1.46.0
netaddr==0.8.0
oauthlib==3.2.2
olefile==0.46
onboard==1.4.1
packaging==24.0
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
proton-core==0.3.3
proton-keyring-linux==0.0.2
proton-keyring-linux-secretservice==0.0.2
proton-vpn-api-core==0.33.11
proton-vpn-gtk-app==4.4.5
proton-vpn-killswitch-network-manager==0.5.4
proton-vpn-killswitch-network-manager-wireguard==0.1.4
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.6.3
psutil==5.9.8
ptyprocess==0.7.0
pulsemixer==1.5.1
pyasyncore==1.0.2
pycairo==1.25.1
pycares==4.4.0
pycryptodomex==3.20.0
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-gnupg==0.5.2
python-xlib==0.33
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
SecretStorage==3.3.3
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
systemd-python==235
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
websockets==10.4
wheel==0.42.0
wxPython==4.2.1
xdg==5
xkit==0.0.0
yarl==1.8.2
yt-dlp==2024.4.9
zipp==1.0.0

# Added for deputados API
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
